python-server/requirements.txt
grpcio==1.18.0
googleapis-common-protos==1.5.6